+++ src/ai/embeddings_request.rs
