+++ src/ai/llm_request.rs
